First framework to reliably pass 80%
While not the first model to reach 80%, this is my first model to have reliably reached and even slightly exceeded the 80% benchmark repeatedly.
The model uses the VGG architecture, with three VGG blocks stacked onto each other, using stochastic gradient descent and cross-entropy loss.  It utilizes three dropout layers to improve the network's overall robustness and accuracy, with a dropout rate of 20%.

Subsequent modifications will include different optimizers and loss functions, as well as data augmentation.  An initial augmentation pipeline was implemented, but a data leak ensued and caused the network's accuracy to reach 100% in training and validation after a single epoch - an impossible circumstance given both the current state-of-the-art accuracy (99.3%) and simple common sense.