Data augmented VGG framework
This variant of the network adds three features:  Increasing dropout, data augmentation, and a variably decaying learning rate.  The first is relatively simple, with the dropout increasing by 0.1 per block, going from 0.2 to 0.5.  The second. is a simple data augmentation algorithm employing image shifting, by a factor of 10%, in width and/or height, and horizontally flipping any image at random.  Finally, the learning rate will decay by a factor of 0.5 every time the validation loss fails to improve after 5 epochs.  The learning rate decay factor was arrived at after experimentation, as lower values tend to bring the optimization to a screeching halt around 78%.

The optimization algorithm is Nadam, and the model was trained for 100 epochs, arriving at 81.85% accuracy, and comfortably sitting in the 80-81% validation accuracy range for the last 30 epochs of training.  Overfitting was reduced to virtually nil, though the first 60 epochs experienced a considerable amount of validation noise.

The model flattened out around 80%, as was mentioned before.  Subsequent attempts to improve will consist of tweaking the augmentation pipeline, the learning rate decay, and perhaps the optimizer.  The number of epochs may be increased, regularization employed, and the decay rate and patience factor increased.